Refactor address controllers within person to be in their own class
